Merge dev to main (second time)

+++ server/src/services/users/blacklist/delete-from-blacklist.ts
@@ -1,4 +1,9 @@
+import type { EmailInput, UserEmail } from '@looking-for-group/shared';
+import { createElement } from 'react';
+import { pretty, render, toPlainText } from 'react-email';
 import prisma from '#config/prisma.ts';
+import UnbanEmail from '#email-templates/unban-email.ts';
+import { sendEmail } from '#services/mailer.ts';
 import type { ServiceErrorSubset, ServiceSuccessSubset } from '#services/service-outcomes.ts';
 
 type DeleteBlacklistServiceError = ServiceErrorSubset<'INTERNAL_ERROR' | 'NOT_FOUND'>;
@@ -7,14 +12,50 @@ type DeleteBlacklistServiceSuccess = ServiceSuccessSubset<'OK'>;
 //DELETE api/mod/unban-user/{id}
 //unbans a user
 const deleteBlacklistService = async (
-  googleId: string,
+  id: number,
 ): Promise<DeleteBlacklistServiceSuccess | DeleteBlacklistServiceError> => {
   try {
-    await prisma.userBlacklist.delete({
+    const user = await prisma.users.findUnique({
       where: {
-        googleId: googleId,
+        userId: id,
       },
     });
+    if (user === null) return 'NOT_FOUND';
+
+    await prisma.userBlacklist.delete({
+      where: {
+        googleId: user.googleId,
+      },
+    });
+
+    //Send email to user
+    const html = await pretty(
+      await render(
+        createElement(UnbanEmail, {
+          receiverName: {
+            firstName: user.firstName,
+            lastName: user.lastName,
+          },
+        }),
+      ),
+    );
+
+    const text = toPlainText(html);
+
+    const email: EmailInput = {
+      sender: {
+        ritEmail: 'lfg-team@lookingforgrp.com',
+        firstName: 'Looking For Group',
+        lastName: '',
+      } as UserEmail,
+      receiver: user,
+      subject: `You have been unbanned from Looking For Group`,
+      textBody: text,
+      HTMLBody: html,
+    };
+
+    const emailResult = await sendEmail(email);
+    if (emailResult === 'INTERNAL_ERROR') return emailResult;
 
     return 'OK';
   } catch (e) {
